requirements.txt
flask==3.0.0
sentence-transformers==3.0.1
pinecone-client==5.0.1
groq>=0.9.0
PyPDF2==3.0.1
python-docx==1.1.0
cohere==5.11.0
python-dotenv==1.0.0
tiktoken==0.5.2